fix: make everything long enough

+++ kernel/src/console.h
@@ -10,4 +10,4 @@ void console_puts(const char *str);
 
 void console_put_with_color(const char *str, uint32_t fg);
 
-void to_hex_string(unsigned int num, char *result);
+void to_hex_string(unsigned long long int num, char *result);
